requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
astroid==2.5.1
    # via pylint
autopep8==1.5.5
    # via -r requirements.in
colorama==0.4.4
    # via pylint
isort==5.7.0
    # via pylint
lazy-object-proxy==1.5.2
    # via astroid
mccabe==0.6.1
    # via pylint
pycodestyle==2.6.0
    # via autopep8
pylint==2.7.2
    # via -r requirements.in
python-dotenv==0.15.0
    # via -r requirements.in
slack-bolt==1.4.3
    # via -r requirements.in
slack-sdk==3.4.2
    # via slack-bolt
toml==0.10.2
    # via
    #   autopep8
    #   pylint
wrapt==1.12.1
    # via astroid
